feat: 반려동물 삭제 페이지 데이터 연동
- TODO: 반려동물 삭제 서버 반영

+++ UMC-Kitten-iOS/Data/Dto/Mypage/MypageResponseDto.swift
@@ -10,14 +10,14 @@ import Foundation
 struct MypageResponseDto {
     
     struct UserResponseDto: Codable {
-        let id: Int
+        let id: Int64
         let nickname: String
         let profileImage: String?
         let hasPet: Bool
         let pets: [PetDto]
         
         struct PetDto: Codable {
-            let id: Int
+            let id: Int64
             let type: PetTypeDto
             let name: String
             let petProfileImage: String
